+++ CHANGES.rst
@@ -9,6 +9,7 @@ WIP
 - `organization owners <https://github.com/openwisp/openwisp-users#organization-owners>`_:
   [admin] Added support for organization owners
 - [admin] Made organization owners read-only to non-superusers
+- [model] Added a helper to cache user permissions
 
 Version 0.2.2 [2020-05-04]
 --------------------------